Improve date option value parsing
Some date strings are parsed as UTC (e.g. YYYY-MM-DD) which should not
have a timezone offset applied while others should (unspecified, but
currently any other date format).

Require value to be a date at midnight (i.e. without time) before or
after conversion from local time to UTC to handle both cases.  This is
not ideal, since dates with times are misinterpreted if the time is the
same as the current timezone offset, but I can't think of anything
better at the moment.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>