Throw TypeError for string testCmdWithArgs
Treating 'string' as Iterable<string> is almost certainly wrong.  Help
the user by failing early with an understandable Error in this case.

Note: We could wrap the string into a single-element Array, but I am
concerned this creates an expectation that the string would be split or
otherwise interpreted by the shell, which we do not do.  Require Array
for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>