feat: use date from version for ordering
I now suspect that the date field from the JSON is the build artifact
directory mtime, which is not a reliable indicator of the build date.
Use the date from the version string, which I expect to be more
reliable.

For an problematic example, see v10.0.0-nightly2018033083d44bee01 which
has date 2018-04-01.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>